Minor improvements for perf script flamegraph
When processing the perf data file generated with multiple events,
the flamegraph script will count all the events regardless of
different event names. If specify the perf data file with -i option,
the script will try to read the header information regardless of
the file name specified, instead it will try to access the perf.data.

This patch tries to add a -e option to specify the event name that
the flamegraph will be generated accordingly. If the -e option omitted,
the behavior remains unchanged. If the -i option specified, the header
information will be read from that file.

Signed-off-by: Tianyou Li <tianyou.li@intel.com>
Reviewed-by: Pan Deng <pan.deng@intel.com>
Reviewed-by: Zhiguo Zhou <zhiguo.zhou@intel.com>
Reviewed-by: Wangyang Guo <wangyang.guo@intel.com>
Reviewed-by: Tim Chen <tim.c.chen@linux.intel.com>