Read perf data file header from the file user specified with -i option for perf script flamegraph
If specify the perf data file with -i option, the script will try to read the header information
regardless of the file name specified, instead it will try to access the perf.data. This simple
patch use the file name from -i option for command perf report --header-only to read the header.

Signed-off-by: Tianyou Li <tianyou.li@intel.com>
Reviewed-by: Pan Deng <pan.deng@intel.com>
Reviewed-by: Zhiguo Zhou <zhiguo.zhou@intel.com>
Reviewed-by: Wangyang Guo <wangyang.guo@intel.com>
Reviewed-by: Tim Chen <tim.c.chen@linux.intel.com>