mm/memory hotplug/unplug: optimize zone->contiguous update when move pfn range
When invoke move_pfn_range_to_zone, it will update the zone->contiguous by
checking the new zone's pfn range from the beginning to the end, regardless
the previous state of the old zone. When the zone's pfn range is large, the
cost of updating the zone->contiguous could be significant.

Add quick path to optimize the zone->contiguous update. When the new range
did not overlap with previous range, the contiguous should be false. If the
new range adjacent with the previous range, just need to check the new
range. If the new added pages could not fill the hole of previous zone, the
contiguous should be false.

Test cases:

Performance diff:

Reported-by: Nanhai Zou <nanhai.zou@intel.com>
Tested-by: Yuan Liu <yuan1.liu@intel.com>
Reviewed-by: Qiuxu Zhuo <qiuxu.zhuo@intel.com>
Reviewed-by: Nanhai Zou <nanhai.zou@intel.com>
Signed-off-by: Tianyou Li <tianyou.li@intel.com>